Add conditional usage of libfuse
