+++ build.rs
@@ -1,4 +1,6 @@
 
 fn main () {
-    println!("cargo:rustc-link-lib=fuse");
+    if !cfg!(feature = "rust-mount") {
+        println!("cargo:rustc-link-lib=fuse");
+    }
 }
