+++ src/channel.rs
@@ -7,7 +7,7 @@ use std::ffi::{CString, CStr, OsStr};
 use std::os::unix::ffi::OsStrExt;
 use std::path::{PathBuf, Path};
 use libc::{self, c_int, c_void, size_t};
-use fuse::fuse_args;
+use fuse::{fuse_args, fuse_mount_compat25, fuse_unmount_compat22};
 use reply::ReplySender;
 
 /// Helper function to provide options as a fuse_args struct
@@ -27,83 +27,6 @@ pub struct Channel {
     fd: c_int,
 }
 
-use libc::getuid;
-use libc::getgid;
-use libc::mount;
-use std::fs::OpenOptions;
-use std::os::unix::io::AsRawFd;
-use std::os::unix::io::IntoRawFd;
-fn fuse_mount_sys(mountpoint: &PathBuf, flags: u64) -> i32
-{
-    // TODO:Check args
-    // TODO:Check mountpoint
-    // TODO:Check nonempty
-    // TODO:Check auto_umount
-    let mut f = OpenOptions::new().read(true).write(true).open("/dev/fuse").unwrap();
-
-
-    // TODO:Check f
-    // from:sdcard.c    sprintf(opts, "fd=%i,rootmode=40000,default_permissions,allow_other,"
-    //                                "user_id=%d,group_id=%d", fd, uid, gid);
-    let opts = format!("fd={},rootmode={},default_permissions,allow_other,user_id={},group_id={}",
-                       f.as_raw_fd(),40000,
-                       unsafe{getuid()}, unsafe{getgid()});
-    // TODO: Add kernel opt
-    // 
-    //TODO: understand:
-    /*
-	strcpy(type, mo->blkdev ? "fuseblk" : "fuse");
-	if (mo->subtype) {
-		strcat(type, ".");
-		strcat(type, mo->subtype);
-	}
-	strcpy(source,
-	       mo->fsname ? mo->fsname : (mo->subtype ? mo->subtype : devname));
-
-    */
-    info!("{}", opts);
-    let c_sources = CString::new("/dev/fuse").unwrap();
-    let c_mnt = CString::new(mountpoint.as_os_str().as_bytes()).unwrap();
-    let c_fs = CString::new("fuse").unwrap();
-    let c_opts = CString::new(opts).unwrap();
-    info!("MOUNT({:?} {:?} {:?} {:?})", c_sources, c_mnt, c_fs, c_opts);
-    let res = unsafe{
-        #[cfg(target_pointer_width = "32")]
-        let flags = flags as u32;
-        mount(c_sources.as_ptr(), c_mnt.as_ptr(), c_fs.as_ptr(), flags, c_opts.as_ptr() as *mut c_void)
-    };
-    if res < 0 {
-        // TODO: error    
-        error!("fuse_mount_sys ERROR: {}", res);
-        panic!("fuse_mount_sys panic!");
-    }
-    f.into_raw_fd()
-}
-
-// /usr/include/sys/mount.h
-const MS_NOSUID :u64 = 2;
-const MS_NODEV  :u64 = 4;
-
-use libc::perror;
-fn fuse_mount_compat25(mountpoint: &PathBuf, args: &fuse_args) -> i32
-{
-    let flags = MS_NOSUID | MS_NODEV;
-
-// TODO: parse fuse_args
-/*
-    pub argc: c_int,
-    pub argv: *const *const c_char,
-    pub allocated: c_int,
-*/
-// TODO: check if allow_other and allow_root aren't mutually active
-// TODO: check if help
-// TODO: get kernel/other flags options
-   let res = fuse_mount_sys(mountpoint, flags);
-   println!("fantafs: fuse_mount_compat25: fd={}", res);
-   // TODO: ERROR
-   res
-}
-
 impl Channel {
     /// Create a new communication channel to the kernel driver by mounting the
     /// given path. The kernel driver will delegate filesystem operations of
@@ -112,12 +35,8 @@ impl Channel {
     pub fn new (mountpoint: &Path, options: &[&OsStr]) -> io::Result<Channel> {
         let mountpoint = try!(mountpoint.canonicalize());
         with_fuse_args(options, |args| {
-            let fd = fuse_mount_compat25(&mountpoint, args);
-            if fd < 0 {
-                Err(io::Error::last_os_error())
-            } else {
-                Ok(Channel { mountpoint: mountpoint, fd: fd })
-            }
+            let fd = fuse_mount_compat25(&mountpoint, args)?;
+            Ok(Channel { mountpoint: mountpoint, fd: fd })
         })
     }
